trimmed-requirements.txt
Flask==2.2.5
Jinja2==3.1.4
MarkupSafe==2.1.5
Werkzeug==2.2.3
SQLAlchemy==1.4.54
ffmpeg-python==0.2.0
moviepy==1.0.3
opencv-python-headless==4.10.0.84
imageio==2.35.1
imageio-ffmpeg==0.5.1
pydub==0.25.1
transformers>=4.44.2
accelerate
safetensors==0.4.5
tqdm==4.66.5
tokenizers==0.19.1
tiktoken==0.7.0
huggingface-hub==0.24.6
fsspec==2024.9.0
filelock==3.16.0
regex==2024.9.11
numpy==2.0.2
scipy==1.14.1
scikit-image==0.24.0
tifffile==2024.8.30
pandas==2.2.3
python-docx==1.1.2
reportlab==4.2.2
pillow==10.4.0
kornia==0.8.1
requests==2.32.3
PyYAML==6.0.2
pytz==2024.2
tzdata==2024.1
python-dateutil==2.9.0.post0
pytest==7.4.3
GitPython==3.1.41
